clean

+++ Sources/CUPS/CUPSPage.swift
@@ -29,18 +29,18 @@ public struct CUPSPage {
     
     public var header: cups_page_header2_t
     
-    public init?(_ media: UnsafePointer<pwg_media_t>, _ type: UnsafePointer<Int8>, _ xdpi: Int32, _ ydpi: Int32, _ sides: UnsafePointer<Int8>?, _ sheet_back: UnsafePointer<Int8>?) {
+    public init?(_ media: UnsafePointer<pwg_media_t>, _ type: UnsafePointer<Int8>, _ xdpi: UInt32, _ ydpi: UInt32, _ sides: UnsafePointer<Int8>?, _ sheet_back: UnsafePointer<Int8>?) {
         self.header = cups_page_header2_t()
-        guard cupsRasterInitPWGHeader(&header, UnsafeMutablePointer(mutating: media), type, xdpi, ydpi, sides, sheet_back) != 0 else { return nil }
+        guard cupsRasterInitPWGHeader(&header, UnsafeMutablePointer(mutating: media), type, Int32(xdpi), Int32(ydpi), sides, sheet_back) != 0 else { return nil }
     }
     
-    public init?(_ media: CUPSMedia, _ type: UnsafePointer<Int8>, _ xdpi: Int32, _ ydpi: Int32, _ sides: UnsafePointer<Int8>?, _ sheet_back: UnsafePointer<Int8>?) {
+    public init?(_ media: CUPSMedia, _ type: UnsafePointer<Int8>, _ xdpi: UInt32, _ ydpi: UInt32, _ sides: UnsafePointer<Int8>?, _ sheet_back: UnsafePointer<Int8>?) {
         guard let page = media.withUnsafePwgMediaPointer(callback: { $0.flatMap { CUPSPage($0, type, xdpi, ydpi, sides, sheet_back) } }) else { return nil }
         self = page
         self.header.Margins.0 = 72 * UInt32(media.left) / 2540
         self.header.Margins.1 = 72 * UInt32(media.bottom) / 2540
-        self.header.cupsWidth = UInt32((media.width - media.left - media.right) * xdpi) / 2540
-        self.header.cupsHeight = UInt32((media.height - media.top - media.bottom) * ydpi) / 2540
+        self.header.cupsWidth = UInt32(media.width - media.left - media.right) * xdpi / 2540
+        self.header.cupsHeight = UInt32(media.height - media.top - media.bottom) * ydpi / 2540
         self.header.cupsBytesPerLine = (self.header.cupsWidth * self.header.cupsBitsPerPixel + 7) / 8
     }
 }
